添加角色完成：向角色表添加一条记录后，通过mybatis返回的主键向角色-菜单，角色-部门，角色-区域映射表中添加数据，并通过spring的声明式事务对这一批操作数据库的动作进行控制

+++ drp/src/main/java/com/lyu/drp/sysmanage/action/RoleAction.java
@@ -89,9 +89,9 @@ public class RoleAction {
 		
 		Role role = JSON.parseObject(jsonObj, Role.class);
 		
-		List<Integer> menuIds = JSON.parseArray(this.menuIds, Integer.class);
-		List<Integer> deptIds = JSON.parseArray(this.deptIds, Integer.class);
-		List<Integer> areaIds = JSON.parseArray(this.areaIds, Integer.class);
+		List<Long> menuIds = JSON.parseArray(this.menuIds, Long.class);
+		List<Long> deptIds = JSON.parseArray(this.deptIds, Long.class);
+		List<Long> areaIds = JSON.parseArray(this.areaIds, Long.class);
 		
 		if (role.getId() == null) { // 新增
 			System.out.println("这是新增--");
